Add simple linter using only control dependence
The linter only implements one check (which is just in the main Run()
function) which checks for non-uniform derivatives. It assumes that all
branch conditions are non-uniform and uses control dependence analysis
to find blocks that depend on them. It outputs a warning for all
derivatives found in non-uniform blocks.